# frontend/openapi.json
{
  "components": {
    "schemas": {
      "CommentModel": {
        "properties": {
          "author": {
            "default": "Anonymous viewer",
            "title": "Author",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "episode_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Code"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mine": {
            "default": false,
            "title": "Mine",
            "type": "boolean"
          },
          "series_id": {
            "title": "Series Id",
            "type": "integer"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "target",
          "series_id",
          "text",
          "created_at"
        ],
        "title": "CommentModel",
        "type": "object"
      },
      "CommentRequest": {
        "properties": {
          "episode_id": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text"
        ],
        "title": "CommentRequest",
        "type": "object"
      },
...
      "EpisodeModel": {
        "properties": {
          "airdate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Airdate"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "runtime_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Minutes"
          },
          "season": {
            "title": "Season",
            "type": "integer"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "watched": {
            "default": false,
            "title": "Watched",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "season",
          "number",
          "code",
          "summary"
        ],
        "title": "EpisodeModel",
        "type": "object"
      },
      "ErrorModel": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorModel",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InsightsResponse": {
        "properties": {
          "based_on_comment_count": {
            "default": 0,
            "title": "Based On Comment Count",
            "type": "integer"
          },
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "degraded": {
            "default": false,
            "title": "Degraded",
            "type": "boolean"
          },
          "generated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "target": {
            "default": "series",
            "title": "Target",
            "type": "string"
          },
          "target_id": {
            "default": 0,
            "title": "Target Id",
            "type": "integer"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "provider"
        ],
        "title": "InsightsResponse",
        "type": "object"
      },
...
      "SearchResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SeriesCardModel"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "query",
          "count"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
      "SeasonModel": {
        "properties": {
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/EpisodeModel"
            },
            "title": "Episodes",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "number": {
            "title": "Number",
            "type": "integer"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "watched": {
            "default": 0,
            "title": "Watched",
            "type": "integer"
          }
        },
        "required": [
          "number",
          "label"
        ],
        "title": "SeasonModel",
        "type": "object"
      },
      "SeriesCardModel": {
        "properties": {
          "genres": {
            "items": {
              "type": "string"
            },
            "title": "Genres",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "network": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Network"
          },
          "poster_thumbnail_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Thumbnail Url"
          },
          "poster_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Poster Url"
          },
          "rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "SeriesCardModel",
        "type": "object"
      },
      "SeriesDetailModel": {
        "properties": {
          "comment_count": {
            "default": 0,
            "title": "Comment Count",
            "type": "integer"
          },
          "comments": {
            "items": {
              "$ref": "#/components/schemas/CommentModel"
            },
            "title": "Comments",
            "type": "array"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "seasons": {
            "items": {
              "$ref": "#/components/schemas/SeasonModel"
            },
            "title": "Seasons",
            "type": "array"
          },
          "series": {
            "$ref": "#/components/schemas/SeriesCardModel"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "total_episodes": {
            "default": 0,
            "title": "Total Episodes",
            "type": "integer"
          },
          "watched": {
            "default": 0,
            "title": "Watched",
            "type": "integer"
          }
        },
        "required": [
          "series",
          "summary"
        ],
        "title": "SeriesDetailModel",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Interactive TV series experience: search, episode tracking, comments and AI powered insights over the public TVMaze API.",
    "title": "tv-insight",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/series/search": {
      "get": {
        "operationId": "search_series_api_series_search_get",
        "parameters": [
          {
            "description": "Free text query",
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "description": "Free text query",
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Search TV series",
        "tags": [
          "series"
        ]
      }
    },
    "/api/series/{series_id}": {
      "get": {
        "operationId": "get_series_detail_api_series__series_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDetailModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Series details",
        "tags": [
          "series"
        ]
      }
    },
    "/api/series/{series_id}/comments": {
      "get": {
        "operationId": "list_comments_api_series__series_id__comments_get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "episode_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Episode Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CommentModel"
                  },
                  "title": "Response List Comments Api Series  Series Id  Comments Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "List comments",
        "tags": [
          "comments"
        ]
      },
      "post": {
        "operationId": "add_comment_api_series__series_id__comments_post",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Add a comment",
        "tags": [
          "comments"
        ]
      }
    },
...
    "/api/series/{series_id}/episodes/{episode_id}/insight": {
      "get": {
        "operationId": "episode_insight_api_series__series_id__episodes__episode_id__insight_get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "integer"
            }
          },
          {
            "description": "Bypass the insight cache",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Bypass the insight cache",
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Insight for an episode",
        "tags": [
          "insights"
        ]
      }
    },
...
    "/api/series/{series_id}/insight": {
      "get": {
        "operationId": "series_insight_api_series__series_id__insight_get",
        "parameters": [
          {
            "in": "path",
            "name": "series_id",
            "required": true,
            "schema": {
              "title": "Series Id",
              "type": "integer"
            }
          },
          {
            "description": "Bypass the insight cache",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "Bypass the insight cache",
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InsightsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Invalid input"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Unknown resource"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Upstream catalogue failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "No insight provider available"
          }
        },
        "summary": "Insight for a series",
        "tags": [
          "insights"
        ]
      }
    },
...
  }
}